Changes required for world state

+++ LuaEngine.cpp
@@ -240,7 +240,7 @@ void Eluna::DestroyBindStores()
 
 void Eluna::RunScripts()
 {
-    ELUNA_LOG_DEBUG("[Eluna]: Running scripts for map: %u", boundMapId);
+    ELUNA_LOG_DEBUG("[Eluna]: Running scripts for map: %i", boundMapId);
 
     uint32 oldMSTime = ElunaUtil::GetCurrTime();
     uint32 count = 0;
@@ -256,9 +256,9 @@ void Eluna::RunScripts()
     for (ScriptList::const_iterator it = sElunaLoader->combined_scripts.begin(); it != sElunaLoader->combined_scripts.end(); ++it)
     {
         // check that the script file is either global or meant to be loaded for this map
-        if (it->mapId != -1 && it->mapId != int32(boundMapId))
+        if (it->mapId != -1 && it->mapId != boundMapId)
         {
-            ELUNA_LOG_DEBUG("[Eluna]: `%s` is tagged %i and will not load for map: %u", it->filename.c_str(), it->mapId, boundMapId);
+            ELUNA_LOG_DEBUG("[Eluna]: `%s` is tagged %i and will not load for map: %i", it->filename.c_str(), it->mapId, boundMapId);
             continue;
         }
 
@@ -311,7 +311,7 @@ void Eluna::RunScripts()
     }
     // Stack: package, modules
     lua_pop(L, 2);
-    ELUNA_LOG_INFO("[Eluna]: Executed %u Lua scripts in %u ms for map state %u", count, ElunaUtil::GetTimeDiff(oldMSTime), boundMapId);
+    ELUNA_LOG_INFO("[Eluna]: Executed %u Lua scripts in %u ms for map state %i", count, ElunaUtil::GetTimeDiff(oldMSTime), boundMapId);
 
     OnLuaStateOpen();
 }
